api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Skiball <divye11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
motor = "^3.5.1"
langchain = "^0.2.11"
langchain-ollama = "^0.1.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
langchain-community = "^0.2.10"
pdfminer = "^20191125"
pdfminer-six = "^20240706"
boto3 = "^1.34.153"
fitz = "^0.0.1.dev2"
pymupdf = "^1.24.9"
docx = "^0.2.4"
python-docx = "^1.1.2"

[tool.poetry.scripts]
start = "api.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
